+++ src/main/java/com/example/helloworldmvc/service/UserService.java
@@ -6,5 +6,4 @@ import com.example.helloworldmvc.web.dto.UserRequestDTO;
 public interface UserService {
     TokenListDTO loginGmail(UserRequestDTO.GoogleEmailRequest request);
 
-    String deactivateUser(String userId);
 }
